Removing shuffle-array dependency
Because there is no tag for shuffle-array 0.1.0 (as per https://github.com/pazguille/shuffle-array/issues/8), including it as a dependency in the bower file causes the ``bower install`` command to fail and the ``bower_components`` directory not to be made. We can trade a little precision and elegance for some functionality by adding a separate shuffle-array install step instead of listing it as a bower dependency.